requirements.txt
absl-py==2.2.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==11.0.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
conformer==0.3.2
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
datasets==2.18.0
decorator==4.4.2
deepspeed==0.14.2
diffusers==0.24.0
dill==0.3.8
distro==1.9.0
editdistance==0.8.1
einops==0.6.1
exceptiongroup==1.2.2
facenet-pytorch==2.5.0
fastapi==0.115.6
fastapi-cli==0.0.4
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.2.0
funasr==1.2.6
future==1.0.0
gast==0.6.0
gdown==5.1.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
gradio==5.4.0
gradio_client==1.4.2
groovy==0.1.2
grpcio==1.57.0
grpcio-tools==1.57.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface==0.0.1
huggingface-hub==0.25.1
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.3.1
jaconv==0.4.0
jamo==0.4.1
jax==0.5.3
jaxlib==0.5.3
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2
lightning==2.2.4
lightning-utilities==0.14.2
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
modelscope==1.15.0
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.1
ninja==1.11.1.4
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.0
onnxruntime-gpu==1.18.0
openai==1.77.0
openai-whisper==20231117
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
orjson==3.10.16
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
proglog==0.1.11
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.19.1
pynini==2.1.5
pynndescent==0.5.13
pynvml==12.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytorch-lightning==2.5.1
pytorch-wpe==0.0.1
pytz==2025.2
pyworld==0.3.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
starlette==0.41.3
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorrt-cu12==10.0.1
tensorrt-cu12-bindings==10.0.1
tensorrt-cu12-libs==10.0.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.3.1+cu121
torch-complex==0.4.4
torchaudio==2.3.1+cu121
torchmetrics==1.7.0
torchtyping==0.1.5
torchvision==0.18.1+cu121
tqdm==4.67.1
transformers==4.40.1
triton==2.3.1
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.13.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.30.0
websockets==12.0
Werkzeug==3.1.3
WeTextProcessing==1.0.3
wget==3.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
